requirements.txt
requests
beautifulsoup4
playwright
pandas
openpyxl
pdfplumber
openai
python-dotenv
